Simplify data-table internals
