+++ packages/data-table-mysql/src/lib/adapter.ts
@@ -22,7 +22,9 @@ import type {
   RowDataPacket,
 } from 'mysql2/promise'
 
+import { compileMysqlMigrationOperations as compileMysqlMigrationOperationsFromCompiler } from './migration-compiler.ts'
 import { compileMysqlOperation } from './sql-compiler.ts'
+import { quoteIdentifier } from './sql-compiler-helpers.ts'
 
 /**
  * Mysql adapter configuration.
@@ -84,7 +86,7 @@ export class MysqlDatabaseAdapter implements DatabaseAdapter {
       return [{ text: compiled.text, values: compiled.values }]
     }
 
-    return compileMysqlMigrationOperations(operation)
+    return compileMysqlMigrationOperationsFromCompiler(operation)
   }
 
   /**
@@ -463,38 +465,6 @@ function normalizeInsertId(
   return header.insertId
 }
 
-function quoteIdentifier(value: string): string {
-  return '`' + value.replace(/`/g, '``') + '`'
-}
-
-function quoteTableRef(table: TableRef): string {
-  if (table.schema) {
-    return quoteIdentifier(table.schema) + '.' + quoteIdentifier(table.name)
-  }
-
-  return quoteIdentifier(table.name)
-}
-
-function quoteLiteral(value: unknown): string {
-  if (value === null) {
-    return 'null'
-  }
-
-  if (typeof value === 'number' || typeof value === 'bigint') {
-    return String(value)
-  }
-
-  if (typeof value === 'boolean') {
-    return value ? 'true' : 'false'
-  }
-
-  if (value instanceof Date) {
-    return quoteLiteral(value.toISOString())
-  }
-
-  return "'" + String(value).replace(/'/g, "''") + "'"
-}
-
 function isInsertOperationKind(kind: DataManipulationRequest['operation']['kind']): boolean {
   return kind === 'insert' || kind === 'insertMany' || kind === 'upsert'
 }
@@ -524,433 +494,4 @@ function isDataManipulationOperation(
     operation.kind === 'upsert' ||
     operation.kind === 'raw'
   )
-}
-
-function compileMysqlMigrationOperations(operation: DataMigrationOperation): SqlStatement[] {
-  if (operation.kind === 'raw') {
-    return [{ text: operation.sql.text, values: [...operation.sql.values] }]
-  }
-
-  if (operation.kind === 'createTable') {
-    let columns = Object.keys(operation.columns).map(
-      (columnName) =>
-        quoteIdentifier(columnName) + ' ' + compileMysqlColumn(operation.columns[columnName]),
-    )
-    let constraints: string[] = []
-
-    if (operation.primaryKey) {
-      constraints.push(
-        'constraint ' +
-          quoteIdentifier(operation.primaryKey.name) +
-          ' primary key (' +
-          operation.primaryKey.columns.map((column) => quoteIdentifier(column)).join(', ') +
-          ')',
-      )
-    }
-
-    for (let unique of operation.uniques ?? []) {
-      constraints.push(
-        'constraint ' +
-          quoteIdentifier(unique.name) +
-          ' ' +
-          'unique (' +
-          unique.columns.map((column) => quoteIdentifier(column)).join(', ') +
-          ')',
-      )
-    }
-
-    for (let check of operation.checks ?? []) {
-      constraints.push(
-        'constraint ' + quoteIdentifier(check.name) + ' ' + 'check (' + check.expression + ')',
-      )
-    }
-
-    for (let foreignKey of operation.foreignKeys ?? []) {
-      let clause =
-        'constraint ' +
-        quoteIdentifier(foreignKey.name) +
-        ' ' +
-        'foreign key (' +
-        foreignKey.columns.map((column) => quoteIdentifier(column)).join(', ') +
-        ') references ' +
-        quoteTableRef(foreignKey.references.table) +
-        ' (' +
-        foreignKey.references.columns.map((column) => quoteIdentifier(column)).join(', ') +
-        ')'
-
-      if (foreignKey.onDelete) {
-        clause += ' on delete ' + foreignKey.onDelete
-      }
-
-      if (foreignKey.onUpdate) {
-        clause += ' on update ' + foreignKey.onUpdate
-      }
-
-      constraints.push(clause)
-    }
-
-    let sql =
-      'create table ' +
-      (operation.ifNotExists ? 'if not exists ' : '') +
-      quoteTableRef(operation.table) +
-      ' (' +
-      [...columns, ...constraints].join(', ') +
-      ')'
-
-    let statements: SqlStatement[] = [{ text: sql, values: [] }]
-
-    if (operation.comment) {
-      statements.push({
-        text:
-          'alter table ' +
-          quoteTableRef(operation.table) +
-          ' comment = ' +
-          quoteLiteral(operation.comment),
-        values: [],
-      })
-    }
-
-    return statements
-  }
-
-  if (operation.kind === 'alterTable') {
-    let statements: SqlStatement[] = []
-
-    for (let change of operation.changes) {
-      let sql = 'alter table ' + quoteTableRef(operation.table) + ' '
-
-      if (change.kind === 'addColumn') {
-        sql +=
-          'add column ' +
-          quoteIdentifier(change.column) +
-          ' ' +
-          compileMysqlColumn(change.definition)
-      } else if (change.kind === 'changeColumn') {
-        sql +=
-          'modify column ' +
-          quoteIdentifier(change.column) +
-          ' ' +
-          compileMysqlColumn(change.definition)
-      } else if (change.kind === 'renameColumn') {
-        sql += 'rename column ' + quoteIdentifier(change.from) + ' to ' + quoteIdentifier(change.to)
-      } else if (change.kind === 'dropColumn') {
-        sql += 'drop column ' + quoteIdentifier(change.column)
-      } else if (change.kind === 'addPrimaryKey') {
-        sql +=
-          'add primary key (' +
-          change.constraint.columns.map((column) => quoteIdentifier(column)).join(', ') +
-          ')'
-      } else if (change.kind === 'dropPrimaryKey') {
-        sql += 'drop primary key'
-      } else if (change.kind === 'addUnique') {
-        sql +=
-          'add ' +
-          'constraint ' +
-          quoteIdentifier(change.constraint.name) +
-          ' ' +
-          'unique (' +
-          change.constraint.columns.map((column) => quoteIdentifier(column)).join(', ') +
-          ')'
-      } else if (change.kind === 'dropUnique') {
-        sql += 'drop index ' + quoteIdentifier(change.name)
-      } else if (change.kind === 'addForeignKey') {
-        sql +=
-          'add ' +
-          'constraint ' +
-          quoteIdentifier(change.constraint.name) +
-          ' ' +
-          'foreign key (' +
-          change.constraint.columns.map((column) => quoteIdentifier(column)).join(', ') +
-          ') references ' +
-          quoteTableRef(change.constraint.references.table) +
-          ' (' +
-          change.constraint.references.columns.map((column) => quoteIdentifier(column)).join(', ') +
-          ')'
-      } else if (change.kind === 'dropForeignKey') {
-        sql += 'drop foreign key ' + quoteIdentifier(change.name)
-      } else if (change.kind === 'addCheck') {
-        sql +=
-          'add ' +
-          'constraint ' +
-          quoteIdentifier(change.constraint.name) +
-          ' ' +
-          'check (' +
-          change.constraint.expression +
-          ')'
-      } else if (change.kind === 'dropCheck') {
-        sql += 'drop check ' + quoteIdentifier(change.name)
-      } else if (change.kind === 'setTableComment') {
-        sql += 'comment = ' + quoteLiteral(change.comment)
-      } else {
-        continue
-      }
-
-      statements.push({ text: sql, values: [] })
-    }
-
-    return statements
-  }
-
-  if (operation.kind === 'renameTable') {
-    return [
-      {
-        text:
-          'rename table ' + quoteTableRef(operation.from) + ' to ' + quoteTableRef(operation.to),
-        values: [],
-      },
-    ]
-  }
-
-  if (operation.kind === 'dropTable') {
-    return [
-      {
-        text:
-          'drop table ' + (operation.ifExists ? 'if exists ' : '') + quoteTableRef(operation.table),
-        values: [],
-      },
-    ]
-  }
-
-  if (operation.kind === 'createIndex') {
-    return [
-      {
-        text:
-          'create ' +
-          (operation.index.unique ? 'unique ' : '') +
-          'index ' +
-          quoteIdentifier(operation.index.name) +
-          ' on ' +
-          quoteTableRef(operation.index.table) +
-          (operation.index.using ? ' using ' + operation.index.using : '') +
-          ' (' +
-          operation.index.columns.map((column) => quoteIdentifier(column)).join(', ') +
-          ')' +
-          (operation.index.where ? ' where ' + operation.index.where : ''),
-        values: [],
-      },
-    ]
-  }
-
-  if (operation.kind === 'dropIndex') {
-    return [
-      {
-        text:
-          'drop index ' + quoteIdentifier(operation.name) + ' on ' + quoteTableRef(operation.table),
-        values: [],
-      },
-    ]
-  }
-
-  if (operation.kind === 'renameIndex') {
-    return [
-      {
-        text:
-          'alter table ' +
-          quoteTableRef(operation.table) +
-          ' rename index ' +
-          quoteIdentifier(operation.from) +
-          ' to ' +
-          quoteIdentifier(operation.to),
-        values: [],
-      },
-    ]
-  }
-
-  if (operation.kind === 'addForeignKey') {
-    return [
-      {
-        text:
-          'alter table ' +
-          quoteTableRef(operation.table) +
-          ' add ' +
-          'constraint ' +
-          quoteIdentifier(operation.constraint.name) +
-          ' ' +
-          'foreign key (' +
-          operation.constraint.columns.map((column) => quoteIdentifier(column)).join(', ') +
-          ') references ' +
-          quoteTableRef(operation.constraint.references.table) +
-          ' (' +
-          operation.constraint.references.columns
-            .map((column) => quoteIdentifier(column))
-            .join(', ') +
-          ')' +
-          (operation.constraint.onDelete ? ' on delete ' + operation.constraint.onDelete : '') +
-          (operation.constraint.onUpdate ? ' on update ' + operation.constraint.onUpdate : ''),
-        values: [],
-      },
-    ]
-  }
-
-  if (operation.kind === 'dropForeignKey') {
-    return [
-      {
-        text:
-          'alter table ' +
-          quoteTableRef(operation.table) +
-          ' drop foreign key ' +
-          quoteIdentifier(operation.name),
-        values: [],
-      },
-    ]
-  }
-
-  if (operation.kind === 'addCheck') {
-    return [
-      {
-        text:
-          'alter table ' +
-          quoteTableRef(operation.table) +
-          ' add ' +
-          'constraint ' +
-          quoteIdentifier(operation.constraint.name) +
-          ' ' +
-          'check (' +
-          operation.constraint.expression +
-          ')',
-        values: [],
-      },
-    ]
-  }
-
-  if (operation.kind === 'dropCheck') {
-    return [
-      {
-        text:
-          'alter table ' +
-          quoteTableRef(operation.table) +
-          ' drop check ' +
-          quoteIdentifier(operation.name),
-        values: [],
-      },
-    ]
-  }
-
-  throw new Error('Unsupported data migration operation kind')
-}
-
-function compileMysqlColumn(definition: ColumnDefinition): string {
-  let parts = [compileMysqlColumnType(definition)]
-
-  if (definition.nullable === false) {
-    parts.push('not null')
-  }
-
-  if (definition.default) {
-    if (definition.default.kind === 'now') {
-      parts.push('default current_timestamp')
-    } else if (definition.default.kind === 'sql') {
-      parts.push('default ' + definition.default.expression)
-    } else {
-      parts.push('default ' + quoteLiteral(definition.default.value))
-    }
-  }
-
-  if (definition.autoIncrement) {
-    parts.push('auto_increment')
-  }
-
-  if (definition.primaryKey) {
-    parts.push('primary key')
-  }
-
-  if (definition.unique) {
-    parts.push('unique')
-  }
-
-  if (definition.computed) {
-    parts.push('generated always as (' + definition.computed.expression + ')')
-    parts.push(definition.computed.stored ? 'stored' : 'virtual')
-  }
-
-  if (definition.references) {
-    let clause =
-      'references ' +
-      quoteTableRef(definition.references.table) +
-      ' (' +
-      definition.references.columns.map((column) => quoteIdentifier(column)).join(', ') +
-      ')'
-
-    if (definition.references.onDelete) {
-      clause += ' on delete ' + definition.references.onDelete
-    }
-
-    if (definition.references.onUpdate) {
-      clause += ' on update ' + definition.references.onUpdate
-    }
-
-    parts.push(clause)
-  }
-
-  if (definition.checks && definition.checks.length > 0) {
-    for (let check of definition.checks) {
-      parts.push('check (' + check.expression + ')')
-    }
-  }
-
-  return parts.join(' ')
-}
-
-function compileMysqlColumnType(definition: ColumnDefinition): string {
-  if (definition.type === 'varchar') {
-    return 'varchar(' + String(definition.length ?? 255) + ')'
-  }
-
-  if (definition.type === 'text') {
-    return 'text'
-  }
-
-  if (definition.type === 'integer') {
-    return definition.unsigned ? 'int unsigned' : 'int'
-  }
-
-  if (definition.type === 'bigint') {
-    return definition.unsigned ? 'bigint unsigned' : 'bigint'
-  }
-
-  if (definition.type === 'decimal') {
-    if (definition.precision !== undefined && definition.scale !== undefined) {
-      return 'decimal(' + String(definition.precision) + ', ' + String(definition.scale) + ')'
-    }
-
-    return 'decimal'
-  }
-
-  if (definition.type === 'boolean') {
-    return 'boolean'
-  }
-
-  if (definition.type === 'uuid') {
-    return 'char(36)'
-  }
-
-  if (definition.type === 'date') {
-    return 'date'
-  }
-
-  if (definition.type === 'time') {
-    return 'time'
-  }
-
-  if (definition.type === 'timestamp') {
-    return 'timestamp'
-  }
-
-  if (definition.type === 'json') {
-    return 'json'
-  }
-
-  if (definition.type === 'binary') {
-    return 'blob'
-  }
-
-  if (definition.type === 'enum') {
-    if (definition.enumValues && definition.enumValues.length > 0) {
-      return 'enum(' + definition.enumValues.map((value) => quoteLiteral(value)).join(', ') + ')'
-    }
-
-    return 'text'
-  }
-
-  return 'text'
 }
