+++ packages/data-table-mysql/src/lib/sql-compiler.ts
@@ -1,17 +1,23 @@
 import { getTableName, getTablePrimaryKey } from '@remix-run/data-table/adapter'
-import type {
-  DataManipulationOperation,
-  Predicate,
-  SqlStatement,
-} from '@remix-run/data-table/adapter'
+import type { DataManipulationOperation, SqlStatement } from '@remix-run/data-table/adapter'
 
-type JoinClause = Extract<DataManipulationOperation, { kind: 'select' }>['joins'][number]
+import {
+  collectColumns,
+  compileFromClause,
+  compileGroupByClause,
+  compileHavingClause,
+  compileLimitClause,
+  compileOffsetClause,
+  compileOrderByClause,
+  compileWhereClause,
+  pushValue,
+  quoteIdentifier,
+  quotePath,
+  type CompileContext,
+  type OperationTable,
+} from './sql-compiler-helpers.ts'
+
 type UpsertOperation = Extract<DataManipulationOperation, { kind: 'upsert' }>
-type OperationTable = Extract<DataManipulationOperation, { kind: 'select' }>['table']
-
-type CompileContext = {
-  values: unknown[]
-}
 
 export function compileMysqlOperation(operation: DataManipulationOperation): SqlStatement {
   if (operation.kind === 'raw') {
@@ -213,259 +219,4 @@ function compileUpsertOperation(operation: UpsertOperation, context: CompileCont
       onDuplicate,
     values: context.values,
   }
-}
-
-function compileFromClause(
-  table: OperationTable,
-  joins: JoinClause[],
-  context: CompileContext,
-): string {
-  let output = ' from ' + quotePath(getTableName(table))
-
-  for (let join of joins) {
-    output +=
-      ' ' +
-      normalizeJoinType(join.type) +
-      ' join ' +
-      quotePath(getTableName(join.table)) +
-      ' on ' +
-      compilePredicate(join.on, context)
-  }
-
-  return output
-}
-
-function compileWhereClause(predicates: Predicate[], context: CompileContext): string {
-  if (predicates.length === 0) {
-    return ''
-  }
-
-  return (
-    ' where ' +
-    predicates.map((predicate) => '(' + compilePredicate(predicate, context) + ')').join(' and ')
-  )
-}
-
-function compileGroupByClause(columns: string[]): string {
-  if (columns.length === 0) {
-    return ''
-  }
-
-  return ' group by ' + columns.map((column) => quotePath(column)).join(', ')
-}
-
-function compileHavingClause(predicates: Predicate[], context: CompileContext): string {
-  if (predicates.length === 0) {
-    return ''
-  }
-
-  return (
-    ' having ' +
-    predicates.map((predicate) => '(' + compilePredicate(predicate, context) + ')').join(' and ')
-  )
-}
-
-function compileOrderByClause(orderBy: { column: string; direction: 'asc' | 'desc' }[]): string {
-  if (orderBy.length === 0) {
-    return ''
-  }
-
-  return (
-    ' order by ' +
-    orderBy
-      .map((clause) => quotePath(clause.column) + ' ' + clause.direction.toUpperCase())
-      .join(', ')
-  )
-}
-
-function compileLimitClause(limit: number | undefined): string {
-  if (limit === undefined) {
-    return ''
-  }
-
-  return ' limit ' + String(limit)
-}
-
-function compileOffsetClause(offset: number | undefined): string {
-  if (offset === undefined) {
-    return ''
-  }
-
-  return ' offset ' + String(offset)
-}
-
-function compilePredicate(predicate: Predicate, context: CompileContext): string {
-  if (predicate.type === 'comparison') {
-    let column = quotePath(predicate.column)
-
-    if (predicate.operator === 'eq') {
-      if (
-        predicate.valueType === 'value' &&
-        (predicate.value === null || predicate.value === undefined)
-      ) {
-        return column + ' is null'
-      }
-
-      let comparisonValue = compileComparisonValue(predicate, context)
-      return column + ' = ' + comparisonValue
-    }
-
-    if (predicate.operator === 'ne') {
-      if (
-        predicate.valueType === 'value' &&
-        (predicate.value === null || predicate.value === undefined)
-      ) {
-        return column + ' is not null'
-      }
-
-      let comparisonValue = compileComparisonValue(predicate, context)
-      return column + ' <> ' + comparisonValue
-    }
-
-    if (predicate.operator === 'gt') {
-      let comparisonValue = compileComparisonValue(predicate, context)
-      return column + ' > ' + comparisonValue
-    }
-
-    if (predicate.operator === 'gte') {
-      let comparisonValue = compileComparisonValue(predicate, context)
-      return column + ' >= ' + comparisonValue
-    }
-
-    if (predicate.operator === 'lt') {
-      let comparisonValue = compileComparisonValue(predicate, context)
-      return column + ' < ' + comparisonValue
-    }
-
-    if (predicate.operator === 'lte') {
-      let comparisonValue = compileComparisonValue(predicate, context)
-      return column + ' <= ' + comparisonValue
-    }
-
-    if (predicate.operator === 'in' || predicate.operator === 'notIn') {
-      let values = Array.isArray(predicate.value) ? predicate.value : []
-
-      if (values.length === 0) {
-        return predicate.operator === 'in' ? '1 = 0' : '1 = 1'
-      }
-
-      let keyword = predicate.operator === 'in' ? 'in' : 'not in'
-
-      return (
-        column +
-        ' ' +
-        keyword +
-        ' (' +
-        values.map((value) => pushValue(context, value)).join(', ') +
-        ')'
-      )
-    }
-
-    if (predicate.operator === 'like') {
-      let comparisonValue = compileComparisonValue(predicate, context)
-      return column + ' like ' + comparisonValue
-    }
-
-    if (predicate.operator === 'ilike') {
-      let comparisonValue = compileComparisonValue(predicate, context)
-      return 'lower(' + column + ') like lower(' + comparisonValue + ')'
-    }
-  }
-
-  if (predicate.type === 'between') {
-    return (
-      quotePath(predicate.column) +
-      ' between ' +
-      pushValue(context, predicate.lower) +
-      ' and ' +
-      pushValue(context, predicate.upper)
-    )
-  }
-
-  if (predicate.type === 'null') {
-    return (
-      quotePath(predicate.column) + (predicate.operator === 'isNull' ? ' is null' : ' is not null')
-    )
-  }
-
-  if (predicate.type === 'logical') {
-    if (predicate.predicates.length === 0) {
-      return predicate.operator === 'and' ? '1 = 1' : '1 = 0'
-    }
-
-    let joiner = predicate.operator === 'and' ? ' and ' : ' or '
-
-    return predicate.predicates
-      .map((child) => '(' + compilePredicate(child, context) + ')')
-      .join(joiner)
-  }
-
-  throw new Error('Unsupported predicate')
-}
-
-function compileComparisonValue(
-  predicate: Extract<Predicate, { type: 'comparison' }>,
-  context: CompileContext,
-): string {
-  if (predicate.valueType === 'column') {
-    return quotePath(predicate.value)
-  }
-
-  return pushValue(context, predicate.value)
-}
-
-function normalizeJoinType(type: string): string {
-  if (type === 'left') {
-    return 'left'
-  }
-
-  if (type === 'right') {
-    return 'right'
-  }
-
-  return 'inner'
-}
-
-function quoteIdentifier(value: string): string {
-  return '`' + value.replace(/`/g, '``') + '`'
-}
-
-function quotePath(path: string): string {
-  if (path === '*') {
-    return '*'
-  }
-
-  return path
-    .split('.')
-    .map((segment) => {
-      if (segment === '*') {
-        return '*'
-      }
-
-      return quoteIdentifier(segment)
-    })
-    .join('.')
-}
-
-function pushValue(context: CompileContext, value: unknown): string {
-  context.values.push(value)
-  return '?'
-}
-
-function collectColumns(rows: Record<string, unknown>[]): string[] {
-  let columns: string[] = []
-  let seen = new Set<string>()
-
-  for (let row of rows) {
-    for (let key in row) {
-      if (!Object.prototype.hasOwnProperty.call(row, key) || seen.has(key)) {
-        continue
-      }
-
-      seen.add(key)
-      columns.push(key)
-    }
-  }
-
-  return columns
 }
