+++ packages/data-table-postgres/src/lib/adapter.ts
@@ -16,7 +16,9 @@ import type {
 import { getTablePrimaryKey } from '@remix-run/data-table/adapter'
 import type { Pool as PostgresPool, PoolClient as PostgresPoolClient } from 'pg'
 
+import { compilePostgresMigrationOperations as compilePostgresMigrationOperationsFromCompiler } from './migration-compiler.ts'
 import { compilePostgresOperation } from './sql-compiler.ts'
+import { quoteIdentifier, quoteTableRef } from './sql-compiler-helpers.ts'
 
 /**
  * Postgres adapter configuration.
@@ -67,12 +69,23 @@ export class PostgresDatabaseAdapter implements DatabaseAdapter {
    * @returns Compiled SQL statements.
    */
   compileSql(operation: DataManipulationOperation | DataMigrationOperation): SqlStatement[] {
-    if (isDataManipulationOperation(operation)) {
-      let compiled = compilePostgresOperation(operation)
-      return [{ text: compiled.text, values: compiled.values }]
-    }
+    switch (operation.kind) {
+      case 'select':
+      case 'count':
+      case 'exists':
+      case 'insert':
+      case 'insertMany':
+      case 'update':
+      case 'delete':
+      case 'upsert':
+      case 'raw': {
+        let compiled = compilePostgresOperation(operation)
+        return [{ text: compiled.text, values: compiled.values }]
+      }
 
-    return compilePostgresMigrationOperations(operation)
+      default:
+        return compilePostgresMigrationOperationsFromCompiler(operation)
+    }
   }
 
   /**
@@ -101,7 +114,7 @@ export class PostgresDatabaseAdapter implements DatabaseAdapter {
     return {
       rows,
       affectedRows: normalizeAffectedRows(request.operation.kind, result.rowCount, rows),
-      insertId: normalizeInsertId(request.operation.kind, request.operation, rows),
+      insertId: normalizeInsertId(request.operation, rows),
     }
   }
 
@@ -130,7 +143,7 @@ export class PostgresDatabaseAdapter implements DatabaseAdapter {
    * @returns `true` when the table exists.
    */
   async hasTable(table: TableRef, transaction?: TransactionToken): Promise<boolean> {
-    let relation = toPostgresRelationName(table)
+    let relation = quoteTableRef(table)
     let client = this.#resolveClient(transaction)
     let result = await client.query('select to_regclass($1) is not null as "exists"', [relation])
     let row = result.rows[0] as Record<string, unknown> | undefined
@@ -149,7 +162,7 @@ export class PostgresDatabaseAdapter implements DatabaseAdapter {
     column: string,
     transaction?: TransactionToken,
   ): Promise<boolean> {
-    let relation = toPostgresRelationName(table)
+    let relation = quoteTableRef(table)
     let client = this.#resolveClient(transaction)
     let result = await client.query(
       'select exists (select 1 from pg_attribute where attrelid = to_regclass($1) and attname = $2 and attnum > 0 and not attisdropped) as "exists"',
@@ -409,11 +422,14 @@ function normalizeAffectedRows(
 }
 
 function normalizeInsertId(
-  kind: DataManipulationRequest['operation']['kind'],
   operation: DataManipulationRequest['operation'],
   rows: Record<string, unknown>[],
 ): unknown {
-  if (!isInsertOperationKind(kind) || !isInsertOperation(operation)) {
+  if (
+    operation.kind !== 'insert' &&
+    operation.kind !== 'insertMany' &&
+    operation.kind !== 'upsert'
+  ) {
     return undefined
   }
 
@@ -427,18 +443,6 @@ function normalizeInsertId(
   let row = rows[rows.length - 1]
 
   return row ? row[key] : undefined
-}
-
-function quoteIdentifier(value: string): string {
-  return '"' + value.replace(/"/g, '""') + '"'
-}
-
-function toPostgresRelationName(table: TableRef): string {
-  if (table.schema) {
-    return quoteIdentifier(table.schema) + '.' + quoteIdentifier(table.name)
-  }
-
-  return quoteIdentifier(table.name)
 }
 
 function toBooleanExists(value: unknown): boolean {
@@ -455,503 +459,4 @@ function toBooleanExists(value: unknown): boolean {
   }
 
   return false
-}
-
-function isInsertOperationKind(kind: DataManipulationRequest['operation']['kind']): boolean {
-  return kind === 'insert' || kind === 'insertMany' || kind === 'upsert'
-}
-
-function isInsertOperation(
-  operation: DataManipulationRequest['operation'],
-): operation is Extract<
-  DataManipulationRequest['operation'],
-  { kind: 'insert' | 'insertMany' | 'upsert' }
-> {
-  return (
-    operation.kind === 'insert' || operation.kind === 'insertMany' || operation.kind === 'upsert'
-  )
-}
-
-function isDataManipulationOperation(
-  operation: DataManipulationOperation | DataMigrationOperation,
-): operation is DataManipulationOperation {
-  return (
-    operation.kind === 'select' ||
-    operation.kind === 'count' ||
-    operation.kind === 'exists' ||
-    operation.kind === 'insert' ||
-    operation.kind === 'insertMany' ||
-    operation.kind === 'update' ||
-    operation.kind === 'delete' ||
-    operation.kind === 'upsert' ||
-    operation.kind === 'raw'
-  )
-}
-
-function compilePostgresMigrationOperations(operation: DataMigrationOperation): SqlStatement[] {
-  if (operation.kind === 'raw') {
-    return [{ text: operation.sql.text, values: [...operation.sql.values] }]
-  }
-
-  if (operation.kind === 'createTable') {
-    let columns = Object.keys(operation.columns).map(
-      (columnName) =>
-        quoteIdentifier(columnName) + ' ' + compilePostgresColumn(operation.columns[columnName]),
-    )
-    let tableConstraints: string[] = []
-
-    if (operation.primaryKey) {
-      tableConstraints.push(
-        'constraint ' +
-          quoteIdentifier(operation.primaryKey.name) +
-          ' primary key (' +
-          operation.primaryKey.columns.map((column) => quoteIdentifier(column)).join(', ') +
-          ')',
-      )
-    }
-
-    for (let unique of operation.uniques ?? []) {
-      tableConstraints.push(
-        'constraint ' +
-          quoteIdentifier(unique.name) +
-          ' ' +
-          'unique (' +
-          unique.columns.map((column) => quoteIdentifier(column)).join(', ') +
-          ')',
-      )
-    }
-
-    for (let check of operation.checks ?? []) {
-      tableConstraints.push(
-        'constraint ' + quoteIdentifier(check.name) + ' ' + 'check (' + check.expression + ')',
-      )
-    }
-
-    for (let foreignKey of operation.foreignKeys ?? []) {
-      let clause =
-        'constraint ' +
-        quoteIdentifier(foreignKey.name) +
-        ' ' +
-        'foreign key (' +
-        foreignKey.columns.map((column) => quoteIdentifier(column)).join(', ') +
-        ') references ' +
-        quoteTableRef(foreignKey.references.table) +
-        ' (' +
-        foreignKey.references.columns.map((column) => quoteIdentifier(column)).join(', ') +
-        ')'
-
-      if (foreignKey.onDelete) {
-        clause += ' on delete ' + foreignKey.onDelete
-      }
-
-      if (foreignKey.onUpdate) {
-        clause += ' on update ' + foreignKey.onUpdate
-      }
-
-      tableConstraints.push(clause)
-    }
-
-    let sql =
-      'create table ' +
-      (operation.ifNotExists ? 'if not exists ' : '') +
-      quoteTableRef(operation.table) +
-      ' (' +
-      [...columns, ...tableConstraints].join(', ') +
-      ')'
-    let statements: SqlStatement[] = [{ text: sql, values: [] }]
-
-    if (operation.comment) {
-      statements.push({
-        text:
-          'comment on table ' +
-          quoteTableRef(operation.table) +
-          ' is ' +
-          quoteLiteral(operation.comment),
-        values: [],
-      })
-    }
-
-    return statements
-  }
-
-  if (operation.kind === 'alterTable') {
-    let sqlStatements: SqlStatement[] = []
-
-    for (let change of operation.changes) {
-      let sql = 'alter table ' + quoteTableRef(operation.table) + ' '
-
-      if (change.kind === 'addColumn') {
-        sql +=
-          'add column ' +
-          quoteIdentifier(change.column) +
-          ' ' +
-          compilePostgresColumn(change.definition)
-      } else if (change.kind === 'changeColumn') {
-        let typeSql = compilePostgresColumnType(change.definition)
-        sql += 'alter column ' + quoteIdentifier(change.column) + ' type ' + typeSql
-      } else if (change.kind === 'renameColumn') {
-        sql += 'rename column ' + quoteIdentifier(change.from) + ' to ' + quoteIdentifier(change.to)
-      } else if (change.kind === 'dropColumn') {
-        sql +=
-          'drop column ' + (change.ifExists ? 'if exists ' : '') + quoteIdentifier(change.column)
-      } else if (change.kind === 'addPrimaryKey') {
-        sql +=
-          'add ' +
-          'constraint ' +
-          quoteIdentifier(change.constraint.name) +
-          ' ' +
-          'primary key (' +
-          change.constraint.columns.map((column) => quoteIdentifier(column)).join(', ') +
-          ')'
-      } else if (change.kind === 'dropPrimaryKey') {
-        sql += 'drop constraint ' + quoteIdentifier(change.name)
-      } else if (change.kind === 'addUnique') {
-        sql +=
-          'add ' +
-          'constraint ' +
-          quoteIdentifier(change.constraint.name) +
-          ' ' +
-          'unique (' +
-          change.constraint.columns.map((column) => quoteIdentifier(column)).join(', ') +
-          ')'
-      } else if (change.kind === 'dropUnique') {
-        sql += 'drop constraint ' + quoteIdentifier(change.name)
-      } else if (change.kind === 'addForeignKey') {
-        sql +=
-          'add ' +
-          'constraint ' +
-          quoteIdentifier(change.constraint.name) +
-          ' ' +
-          'foreign key (' +
-          change.constraint.columns.map((column) => quoteIdentifier(column)).join(', ') +
-          ') references ' +
-          quoteTableRef(change.constraint.references.table) +
-          ' (' +
-          change.constraint.references.columns.map((column) => quoteIdentifier(column)).join(', ') +
-          ')'
-      } else if (change.kind === 'dropForeignKey') {
-        sql += 'drop constraint ' + quoteIdentifier(change.name)
-      } else if (change.kind === 'addCheck') {
-        sql +=
-          'add ' +
-          'constraint ' +
-          quoteIdentifier(change.constraint.name) +
-          ' ' +
-          'check (' +
-          change.constraint.expression +
-          ')'
-      } else if (change.kind === 'dropCheck') {
-        sql += 'drop constraint ' + quoteIdentifier(change.name)
-      } else if (change.kind === 'setTableComment') {
-        sqlStatements.push({
-          text:
-            'comment on table ' +
-            quoteTableRef(operation.table) +
-            ' is ' +
-            quoteLiteral(change.comment),
-          values: [],
-        })
-        continue
-      } else {
-        continue
-      }
-
-      sqlStatements.push({ text: sql, values: [] })
-    }
-
-    return sqlStatements
-  }
-
-  if (operation.kind === 'renameTable') {
-    return [
-      {
-        text:
-          'alter table ' +
-          quoteTableRef(operation.from) +
-          ' rename to ' +
-          quoteIdentifier(operation.to.name),
-        values: [],
-      },
-    ]
-  }
-
-  if (operation.kind === 'dropTable') {
-    return [
-      {
-        text:
-          'drop table ' +
-          (operation.ifExists ? 'if exists ' : '') +
-          quoteTableRef(operation.table) +
-          (operation.cascade ? ' cascade' : ''),
-        values: [],
-      },
-    ]
-  }
-
-  if (operation.kind === 'createIndex') {
-    return [
-      {
-        text:
-          'create ' +
-          (operation.index.unique ? 'unique ' : '') +
-          'index ' +
-          (operation.ifNotExists ? 'if not exists ' : '') +
-          quoteIdentifier(operation.index.name) +
-          ' on ' +
-          quoteTableRef(operation.index.table) +
-          (operation.index.using ? ' using ' + operation.index.using : '') +
-          ' (' +
-          operation.index.columns.map((column) => quoteIdentifier(column)).join(', ') +
-          ')' +
-          (operation.index.where ? ' where ' + operation.index.where : ''),
-        values: [],
-      },
-    ]
-  }
-
-  if (operation.kind === 'dropIndex') {
-    return [
-      {
-        text:
-          'drop index ' +
-          (operation.ifExists ? 'if exists ' : '') +
-          quoteIdentifier(operation.name),
-        values: [],
-      },
-    ]
-  }
-
-  if (operation.kind === 'renameIndex') {
-    return [
-      {
-        text:
-          'alter index ' +
-          quoteIdentifier(operation.from) +
-          ' rename to ' +
-          quoteIdentifier(operation.to),
-        values: [],
-      },
-    ]
-  }
-
-  if (operation.kind === 'addForeignKey') {
-    return [
-      {
-        text:
-          'alter table ' +
-          quoteTableRef(operation.table) +
-          ' add ' +
-          'constraint ' +
-          quoteIdentifier(operation.constraint.name) +
-          ' ' +
-          'foreign key (' +
-          operation.constraint.columns.map((column) => quoteIdentifier(column)).join(', ') +
-          ') references ' +
-          quoteTableRef(operation.constraint.references.table) +
-          ' (' +
-          operation.constraint.references.columns
-            .map((column) => quoteIdentifier(column))
-            .join(', ') +
-          ')' +
-          (operation.constraint.onDelete ? ' on delete ' + operation.constraint.onDelete : '') +
-          (operation.constraint.onUpdate ? ' on update ' + operation.constraint.onUpdate : ''),
-        values: [],
-      },
-    ]
-  }
-
-  if (operation.kind === 'dropForeignKey') {
-    return [
-      {
-        text:
-          'alter table ' +
-          quoteTableRef(operation.table) +
-          ' drop constraint ' +
-          quoteIdentifier(operation.name),
-        values: [],
-      },
-    ]
-  }
-
-  if (operation.kind === 'addCheck') {
-    return [
-      {
-        text:
-          'alter table ' +
-          quoteTableRef(operation.table) +
-          ' add ' +
-          'constraint ' +
-          quoteIdentifier(operation.constraint.name) +
-          ' ' +
-          'check (' +
-          operation.constraint.expression +
-          ')',
-        values: [],
-      },
-    ]
-  }
-
-  if (operation.kind === 'dropCheck') {
-    return [
-      {
-        text:
-          'alter table ' +
-          quoteTableRef(operation.table) +
-          ' drop constraint ' +
-          quoteIdentifier(operation.name),
-        values: [],
-      },
-    ]
-  }
-
-  throw new Error('Unsupported data migration operation kind')
-}
-
-function compilePostgresColumn(definition: ColumnDefinition): string {
-  let parts = [compilePostgresColumnType(definition)]
-
-  if (definition.nullable === false) {
-    parts.push('not null')
-  }
-
-  if (definition.default) {
-    if (definition.default.kind === 'now') {
-      parts.push('default now()')
-    } else if (definition.default.kind === 'sql') {
-      parts.push('default ' + definition.default.expression)
-    } else {
-      parts.push('default ' + quoteLiteral(definition.default.value))
-    }
-  }
-
-  if (definition.primaryKey) {
-    parts.push('primary key')
-  }
-
-  if (definition.unique) {
-    parts.push('unique')
-  }
-
-  if (definition.computed) {
-    if (!definition.computed.stored) {
-      throw new Error('Postgres only supports stored computed/generated columns')
-    }
-
-    parts.push('generated always as (' + definition.computed.expression + ') stored')
-  }
-
-  if (definition.references) {
-    let clause =
-      'references ' +
-      quoteTableRef(definition.references.table) +
-      ' (' +
-      definition.references.columns.map((column) => quoteIdentifier(column)).join(', ') +
-      ')'
-
-    if (definition.references.onDelete) {
-      clause += ' on delete ' + definition.references.onDelete
-    }
-
-    if (definition.references.onUpdate) {
-      clause += ' on update ' + definition.references.onUpdate
-    }
-
-    parts.push(clause)
-  }
-
-  if (definition.checks && definition.checks.length > 0) {
-    for (let check of definition.checks) {
-      parts.push('check (' + check.expression + ')')
-    }
-  }
-
-  return parts.join(' ')
-}
-
-function compilePostgresColumnType(definition: ColumnDefinition): string {
-  if (definition.type === 'varchar') {
-    return 'varchar(' + String(definition.length ?? 255) + ')'
-  }
-
-  if (definition.type === 'text') {
-    return 'text'
-  }
-
-  if (definition.type === 'integer') {
-    return 'integer'
-  }
-
-  if (definition.type === 'bigint') {
-    return 'bigint'
-  }
-
-  if (definition.type === 'decimal') {
-    if (definition.precision !== undefined && definition.scale !== undefined) {
-      return 'decimal(' + String(definition.precision) + ', ' + String(definition.scale) + ')'
-    }
-
-    return 'decimal'
-  }
-
-  if (definition.type === 'boolean') {
-    return 'boolean'
-  }
-
-  if (definition.type === 'uuid') {
-    return 'uuid'
-  }
-
-  if (definition.type === 'date') {
-    return 'date'
-  }
-
-  if (definition.type === 'time') {
-    return definition.withTimezone ? 'time with time zone' : 'time without time zone'
-  }
-
-  if (definition.type === 'timestamp') {
-    return definition.withTimezone ? 'timestamp with time zone' : 'timestamp without time zone'
-  }
-
-  if (definition.type === 'json') {
-    return 'jsonb'
-  }
-
-  if (definition.type === 'binary') {
-    return 'bytea'
-  }
-
-  if (definition.type === 'enum') {
-    return 'text'
-  }
-
-  return 'text'
-}
-
-function quoteTableRef(table: TableRef): string {
-  if (table.schema) {
-    return quoteIdentifier(table.schema) + '.' + quoteIdentifier(table.name)
-  }
-
-  return quoteIdentifier(table.name)
-}
-
-function quoteLiteral(value: unknown): string {
-  if (value === null) {
-    return 'null'
-  }
-
-  if (typeof value === 'number' || typeof value === 'bigint') {
-    return String(value)
-  }
-
-  if (typeof value === 'boolean') {
-    return value ? 'true' : 'false'
-  }
-
-  if (value instanceof Date) {
-    return quoteLiteral(value.toISOString())
-  }
-
-  return "'" + String(value).replace(/'/g, "''") + "'"
 }
