+++ packages/data-table/src/lib/adapter.ts
@@ -1,653 +1,70 @@
-import type { AnyTable, OrderByClause } from './table.ts'
-import type { Predicate } from './operators.ts'
-import type { SqlStatement } from './sql.ts'
-import type { Pretty } from './types.ts'
-
-/**
- * Supported SQL join kinds.
- */
-export type JoinType = 'inner' | 'left' | 'right'
-
-/**
- * Join configuration used in compiled select statements.
- */
-export type JoinClause = {
-  type: JoinType
-  table: AnyTable
-  on: Predicate
-}
-
-/**
- * Selected output column with optional alias.
- */
-export type SelectColumn = {
-  column: string
-  alias: string
-}
-
-/**
- * Returning selection for write statements.
- */
-export type ReturningSelection = '*' | string[]
-
-/**
- * Canonical select statement shape consumed by adapters.
- */
-export type SelectOperation<table extends AnyTable = AnyTable> = {
-  kind: 'select'
-  table: table
-  select: '*' | SelectColumn[]
-  distinct: boolean
-  joins: JoinClause[]
-  where: Predicate[]
-  groupBy: string[]
-  having: Predicate[]
-  orderBy: OrderByClause[]
-  limit?: number
-  offset?: number
-}
-
-/**
- * Canonical count statement shape consumed by adapters.
- */
-export type CountOperation<table extends AnyTable = AnyTable> = {
-  kind: 'count'
-  table: table
-  joins: JoinClause[]
-  where: Predicate[]
-  groupBy: string[]
-  having: Predicate[]
-}
-
-/**
- * Canonical exists statement shape consumed by adapters.
- */
-export type ExistsOperation<table extends AnyTable = AnyTable> = {
-  kind: 'exists'
-  table: table
-  joins: JoinClause[]
-  where: Predicate[]
-  groupBy: string[]
-  having: Predicate[]
-}
-
-/**
- * Canonical insert statement shape consumed by adapters.
- */
-export type InsertOperation<table extends AnyTable = AnyTable> = {
-  kind: 'insert'
-  table: table
-  values: Record<string, unknown>
-  returning?: ReturningSelection
-}
-
-/**
- * Canonical bulk-insert statement shape consumed by adapters.
- */
-export type InsertManyOperation<table extends AnyTable = AnyTable> = {
-  kind: 'insertMany'
-  table: table
-  values: Record<string, unknown>[]
-  returning?: ReturningSelection
-}
-
-/**
- * Canonical update statement shape consumed by adapters.
- */
-export type UpdateOperation<table extends AnyTable = AnyTable> = {
-  kind: 'update'
-  table: table
-  changes: Record<string, unknown>
-  where: Predicate[]
-  returning?: ReturningSelection
-}
-
-/**
- * Canonical delete statement shape consumed by adapters.
- */
-export type DeleteOperation<table extends AnyTable = AnyTable> = {
-  kind: 'delete'
-  table: table
-  where: Predicate[]
-  returning?: ReturningSelection
-}
-
-/**
- * Canonical upsert statement shape consumed by adapters.
- */
-export type UpsertOperation<table extends AnyTable = AnyTable> = {
-  kind: 'upsert'
-  table: table
-  values: Record<string, unknown>
-  conflictTarget?: string[]
-  update?: Record<string, unknown>
-  returning?: ReturningSelection
-}
-
-/**
- * Raw SQL statement execution descriptor.
- */
-export type RawOperation = {
-  kind: 'raw'
-  sql: SqlStatement
-}
-
-/**
- * Union of all data-manipulation statement shapes.
- */
-export type DataManipulationOperation =
-  | SelectOperation
-  | CountOperation
-  | ExistsOperation
-  | InsertOperation
-  | InsertManyOperation
-  | UpdateOperation
-  | DeleteOperation
-  | UpsertOperation
-  | RawOperation
-
-/**
- * Qualified table reference used in migration operations.
- */
-export type TableRef = {
-  name: string
-  schema?: string
-}
-
-/**
- * Referential actions supported by foreign key constraints.
- */
-export type ForeignKeyAction = 'cascade' | 'restrict' | 'set null' | 'set default' | 'no action'
-
-/**
- * Logical column type names used by schema definitions.
- */
-export type ColumnTypeName =
-  | 'varchar'
-  | 'text'
-  | 'integer'
-  | 'bigint'
-  | 'decimal'
-  | 'boolean'
-  | 'uuid'
-  | 'date'
-  | 'time'
-  | 'timestamp'
-  | 'json'
-  | 'binary'
-  | 'enum'
-
-/**
- * Default value definition for a column.
- */
-export type ColumnDefault =
-  | { kind: 'literal'; value: unknown }
-  | { kind: 'now' }
-  | { kind: 'sql'; expression: string }
-
-/**
- * Definition for a computed or generated column.
- */
-export type ColumnComputed = {
-  expression: string
-  stored: boolean
-}
-
-/** Options for configuring identity column generation. */
-export type IdentityOptions = {
-  always?: boolean
-  start?: number
-  increment?: number
-}
-
-/** Foreign-key reference metadata declared on a column. */
-export type ColumnReference = {
-  table: TableRef
-  columns: string[]
-  name: string
-  onDelete?: ForeignKeyAction
-  onUpdate?: ForeignKeyAction
-}
-
-/**
- * Check constraint declared on a column definition.
- */
-export type ColumnCheck = {
-  expression: string
-  name: string
-}
-
-/**
- * Normalized column definition used in schema operations.
- */
-export type ColumnDefinition = {
-  type: ColumnTypeName
-  nullable?: boolean
-  primaryKey?: boolean
-  unique?: boolean | { name?: string }
-  default?: ColumnDefault
-  computed?: ColumnComputed
-  references?: ColumnReference
-  checks?: ColumnCheck[]
-  comment?: string
-  length?: number
-  precision?: number
-  scale?: number
-  unsigned?: boolean
-  withTimezone?: boolean
-  enumValues?: string[]
-  autoIncrement?: boolean
-  identity?: IdentityOptions
-  collate?: string
-  charset?: string
-}
-
-/**
- * Primary key constraint definition.
- */
-export type PrimaryKeyConstraint = {
-  columns: string[]
-  name: string
-}
-
-/**
- * Unique constraint definition.
- */
-export type UniqueConstraint = {
-  columns: string[]
-  name: string
-}
-
-/**
- * Check constraint definition.
- */
-export type CheckConstraint = {
-  expression: string
-  name: string
-}
-
-/**
- * Foreign key constraint definition.
- */
-export type ForeignKeyConstraint = {
-  columns: string[]
-  references: {
-    table: TableRef
-    columns: string[]
-  }
-  name: string
-  onDelete?: ForeignKeyAction
-  onUpdate?: ForeignKeyAction
-}
-
-/**
- * Index method used when creating an index.
- */
-export type IndexMethod = 'btree' | 'hash' | 'gin' | 'gist' | 'fulltext' | (string & {})
-
-/**
- * Index definition used in schema operations.
- */
-export type IndexDefinition = {
-  table: TableRef
-  name: string
-  columns: string[]
-  unique?: boolean
-  where?: string
-  using?: IndexMethod
-}
-
-/**
- * Operation that creates a new table.
- */
-export type CreateTableOperation = {
-  kind: 'createTable'
-  table: TableRef
-  ifNotExists?: boolean
-  columns: Record<string, ColumnDefinition>
-  primaryKey?: PrimaryKeyConstraint
-  uniques?: UniqueConstraint[]
-  checks?: CheckConstraint[]
-  foreignKeys?: ForeignKeyConstraint[]
-  comment?: string
-}
-
-/**
- * Alter-table change that adds a column.
- */
-export type AddColumnChange = {
-  kind: 'addColumn'
-  column: string
-  definition: ColumnDefinition
-}
-
-/**
- * Alter-table change that replaces a column definition.
- */
-export type ChangeColumnChange = {
-  kind: 'changeColumn'
-  column: string
-  definition: ColumnDefinition
-}
-
-/**
- * Alter-table change that renames a column.
- */
-export type RenameColumnChange = {
-  kind: 'renameColumn'
-  from: string
-  to: string
-}
-
-/**
- * Alter-table change that drops a column.
- */
-export type DropColumnChange = {
-  kind: 'dropColumn'
-  column: string
-  ifExists?: boolean
-}
-
-/**
- * Alter-table change that adds a primary key.
- */
-export type AddPrimaryKeyChange = {
-  kind: 'addPrimaryKey'
-  constraint: PrimaryKeyConstraint
-}
-
-/**
- * Alter-table change that drops a primary key.
- */
-export type DropPrimaryKeyChange = {
-  kind: 'dropPrimaryKey'
-  name: string
-}
-
-/**
- * Alter-table change that adds a unique constraint.
- */
-export type AddUniqueChange = {
-  kind: 'addUnique'
-  constraint: UniqueConstraint
-}
-
-/**
- * Alter-table change that drops a unique constraint.
- */
-export type DropUniqueChange = {
-  kind: 'dropUnique'
-  name: string
-}
-
-/**
- * Alter-table change that adds a foreign key constraint.
- */
-export type AddForeignKeyChange = {
-  kind: 'addForeignKey'
-  constraint: ForeignKeyConstraint
-}
-
-/**
- * Alter-table change that drops a foreign key constraint.
- */
-export type DropForeignKeyChange = {
-  kind: 'dropForeignKey'
-  name: string
-}
-
-/**
- * Alter-table change that adds a check constraint.
- */
-export type AddCheckChange = {
-  kind: 'addCheck'
-  constraint: CheckConstraint
-}
-
-/**
- * Alter-table change that drops a check constraint.
- */
-export type DropCheckChange = {
-  kind: 'dropCheck'
-  name: string
-}
-
-/**
- * Alter-table change that updates a table comment.
- */
-export type SetTableCommentChange = {
-  kind: 'setTableComment'
-  comment: string
-}
-
-/**
- * Union of supported `alterTable` changes.
- */
-export type AlterTableChange =
-  | AddColumnChange
-  | ChangeColumnChange
-  | RenameColumnChange
-  | DropColumnChange
-  | AddPrimaryKeyChange
-  | DropPrimaryKeyChange
-  | AddUniqueChange
-  | DropUniqueChange
-  | AddForeignKeyChange
-  | DropForeignKeyChange
-  | AddCheckChange
-  | DropCheckChange
-  | SetTableCommentChange
-
-/**
- * Operation that applies one or more table changes.
- */
-export type AlterTableOperation = {
-  kind: 'alterTable'
-  table: TableRef
-  changes: AlterTableChange[]
-  ifExists?: boolean
-}
-
-/**
- * Operation that renames a table.
- */
-export type RenameTableOperation = {
-  kind: 'renameTable'
-  from: TableRef
-  to: TableRef
-}
-
-/**
- * Operation that drops a table.
- */
-export type DropTableOperation = {
-  kind: 'dropTable'
-  table: TableRef
-  ifExists?: boolean
-  cascade?: boolean
-}
-
-/**
- * Operation that creates an index.
- */
-export type CreateIndexOperation = {
-  kind: 'createIndex'
-  index: IndexDefinition
-  ifNotExists?: boolean
-}
-
-/**
- * Operation that drops an index.
- */
-export type DropIndexOperation = {
-  kind: 'dropIndex'
-  table: TableRef
-  name: string
-  ifExists?: boolean
-}
-
-/**
- * Operation that renames an index.
- */
-export type RenameIndexOperation = {
-  kind: 'renameIndex'
-  table: TableRef
-  from: string
-  to: string
-}
-
-/**
- * Operation that adds a table-level foreign key.
- */
-export type AddForeignKeyOperation = {
-  kind: 'addForeignKey'
-  table: TableRef
-  constraint: ForeignKeyConstraint
-}
-
-/**
- * Operation that drops a table-level foreign key.
- */
-export type DropForeignKeyOperation = {
-  kind: 'dropForeignKey'
-  table: TableRef
-  name: string
-}
-
-/**
- * Operation that adds a table-level check constraint.
- */
-export type AddCheckOperation = {
-  kind: 'addCheck'
-  table: TableRef
-  constraint: CheckConstraint
-}
-
-/**
- * Operation that drops a table-level check constraint.
- */
-export type DropCheckOperation = {
-  kind: 'dropCheck'
-  table: TableRef
-  name: string
-}
-
-/**
- * Union of schema and migration operations understood by adapters.
- */
-export type DataMigrationOperation =
-  | CreateTableOperation
-  | AlterTableOperation
-  | RenameTableOperation
-  | DropTableOperation
-  | CreateIndexOperation
-  | DropIndexOperation
-  | RenameIndexOperation
-  | AddForeignKeyOperation
-  | DropForeignKeyOperation
-  | AddCheckOperation
-  | DropCheckOperation
-  | RawOperation
-
-/**
- * Opaque transaction handle supplied by adapters.
- */
-export type TransactionToken = {
-  id: string
-  metadata?: Record<string, unknown>
-}
-
-/**
- * Transaction hints that adapters may apply when supported by the dialect.
- */
-export type TransactionOptions = {
-  isolationLevel?: 'read uncommitted' | 'read committed' | 'repeatable read' | 'serializable'
-  readOnly?: boolean
-}
-
-/**
- * Adapter execution request payload.
- */
-export type DataManipulationRequest = {
-  operation: DataManipulationOperation
-  transaction?: TransactionToken
-}
-
-/**
- * Adapter migration request payload.
- */
-export type DataMigrationRequest = {
-  operation: DataMigrationOperation
-  transaction?: TransactionToken
-}
-
-/**
- * Adapter data-manipulation result payload.
- */
-export type DataManipulationResult = {
-  rows?: Record<string, unknown>[]
-  affectedRows?: number
-  insertId?: unknown
-}
-
-/**
- * Adapter data-migration result payload.
- */
-export type DataMigrationResult = {
-  /**
-   * Number of migration operations processed by the adapter call.
-   */
-  affectedOperations?: number
-}
-
-/**
- * Declares adapter feature support.
- */
-export type AdapterCapabilities = {
-  returning: boolean
-  savepoints: boolean
-  upsert: boolean
-  transactionalDdl: boolean
-  migrationLock: boolean
-}
-
-/**
- * Partial capabilities used to override adapter defaults.
- */
-export type AdapterCapabilityOverrides = Pretty<Partial<AdapterCapabilities>>
-
-/**
- * Runtime contract implemented by concrete database adapters.
- */
-export interface DatabaseAdapter {
-  /** Database dialect name exposed by the adapter. */
-  dialect: string
-  /** Feature flags describing the adapter's supported behaviors. */
-  capabilities: AdapterCapabilities
-  /** Compiles a data or migration operation into executable SQL statements. */
-  compileSql(operation: DataManipulationOperation | DataMigrationOperation): SqlStatement[]
-  /** Executes a data-manipulation request. */
-  execute(request: DataManipulationRequest): Promise<DataManipulationResult>
-  /** Executes a migration request. */
-  migrate(request: DataMigrationRequest): Promise<DataMigrationResult>
-  /** Checks whether a table exists. */
-  hasTable(table: TableRef, transaction?: TransactionToken): Promise<boolean>
-  /** Checks whether a column exists on a table. */
-  hasColumn(table: TableRef, column: string, transaction?: TransactionToken): Promise<boolean>
-  /** Starts a new database transaction. */
-  beginTransaction(options?: TransactionOptions): Promise<TransactionToken>
-  /** Commits an open transaction. */
-  commitTransaction(token: TransactionToken): Promise<void>
-  /** Rolls back an open transaction. */
-  rollbackTransaction(token: TransactionToken): Promise<void>
-  /** Creates a savepoint inside an open transaction. */
-  createSavepoint(token: TransactionToken, name: string): Promise<void>
-  /** Rolls back to a previously created savepoint. */
-  rollbackToSavepoint(token: TransactionToken, name: string): Promise<void>
-  /** Releases a previously created savepoint. */
-  releaseSavepoint(token: TransactionToken, name: string): Promise<void>
-  /** Acquires the adapter's migration lock when supported. */
-  acquireMigrationLock?(): Promise<void>
-  /** Releases the adapter's migration lock when supported. */
-  releaseMigrationLock?(): Promise<void>
-}
+export type {
+  CountOperation,
+  DataManipulationOperation,
+  DeleteOperation,
+  ExistsOperation,
+  InsertManyOperation,
+  InsertOperation,
+  JoinClause,
+  JoinType,
+  RawOperation,
+  ReturningSelection,
+  SelectColumn,
+  SelectOperation,
+  UpdateOperation,
+  UpsertOperation,
+} from './adapter/data-manipulation.ts'
+export type {
+  AddCheckChange,
+  AddCheckOperation,
+  AddColumnChange,
+  AddForeignKeyChange,
+  AddForeignKeyOperation,
+  AddPrimaryKeyChange,
+  AddUniqueChange,
+  AlterTableChange,
+  AlterTableOperation,
+  ChangeColumnChange,
+  CheckConstraint,
+  ColumnCheck,
+  ColumnComputed,
+  ColumnDefault,
+  ColumnDefinition,
+  ColumnReference,
+  ColumnTypeName,
+  CreateIndexOperation,
+  CreateTableOperation,
+  DataMigrationOperation,
+  DropCheckChange,
+  DropCheckOperation,
+  DropColumnChange,
+  DropForeignKeyChange,
+  DropForeignKeyOperation,
+  DropIndexOperation,
+  DropPrimaryKeyChange,
+  DropTableOperation,
+  DropUniqueChange,
+  ForeignKeyAction,
+  ForeignKeyConstraint,
+  IdentityOptions,
+  IndexDefinition,
+  IndexMethod,
+  PrimaryKeyConstraint,
+  RenameColumnChange,
+  RenameIndexOperation,
+  RenameTableOperation,
+  SetTableCommentChange,
+  TableRef,
+  UniqueConstraint,
+} from './adapter/migration.ts'
+export type {
+  AdapterCapabilities,
+  AdapterCapabilityOverrides,
+  DataManipulationRequest,
+  DataManipulationResult,
+  DataMigrationRequest,
+  DataMigrationResult,
+  DatabaseAdapter,
+  TransactionOptions,
+  TransactionToken,
+} from './adapter/runtime.ts'
