+++ packages/data-table/src/lib/query.ts
@@ -1,5 +1,5 @@
-import type { JoinClause, JoinType, SelectColumn } from './adapter.ts'
-import { DataTableQueryError, DataTableValidationError } from './errors.ts'
+import type { JoinType } from './adapter.ts'
+import { DataTableQueryError } from './errors.ts'
 import type {
   MergeColumnTypeMaps,
   PrimaryKeyInputForRow,
@@ -10,86 +10,29 @@ import type {
   QueryColumns,
   QueryTableInput,
   RelationMapForSourceName,
-  ReturningInput,
   SelectedAliasRow,
   WriteResult,
   WriteRowResult,
   WriteRowsResult,
 } from './database.ts'
 import type { Predicate, WhereInput } from './operators.ts'
-import { normalizeWhereInput } from './operators.ts'
 import { normalizeColumnInput } from './references.ts'
-import type { AnyRelation, LoadedRelationMap } from './table-relations.ts'
-import type { AnyTable, OrderByClause } from './table.ts'
-import { getTableColumns, getTableName } from './table.ts'
+import type { LoadedRelationMap } from './table-relations.ts'
+import type { AnyTable } from './table.ts'
+import {
+  assertWriteState,
+  createPredicateColumnResolver,
+  normalizePredicateValues,
+  normalizeQueryWhereInput,
+} from './query/predicate.ts'
+import type { DeleteQueryOptions, InsertQueryOptions, QueryExecutionMode, QueryPlan, UpsertQueryOptions } from './query/plan.ts'
+import { cloneQueryPlan } from './query/plan.ts'
+import type { QueryState } from './query/state.ts'
+import { cloneQueryState, createInitialQueryState, mergeQueryState } from './query/state.ts'
+import type { QuerySnapshot } from './query/snapshot.ts'
+import { createQuerySnapshot } from './query/snapshot.ts'
 
 type QueryBindingState = 'bound' | 'unbound'
-
-type InsertQueryOptions<row extends Record<string, unknown>> = {
-  returning?: ReturningInput<row>
-  touch?: boolean
-}
-
-type DeleteQueryOptions<row extends Record<string, unknown>> = {
-  returning?: ReturningInput<row>
-}
-
-type UpsertQueryOptions<row extends Record<string, unknown>> = {
-  returning?: ReturningInput<row>
-  touch?: boolean
-  conflictTarget?: (keyof row & string)[]
-  update?: Partial<row>
-}
-
-export type QueryState = {
-  select: '*' | SelectColumn[]
-  distinct: boolean
-  joins: JoinClause[]
-  where: Predicate<string>[]
-  groupBy: string[]
-  having: Predicate<string>[]
-  orderBy: OrderByClause[]
-  limit?: number
-  offset?: number
-  with: Record<string, AnyRelation>
-}
-
-type QueryPlanMap<row extends Record<string, unknown>, primaryKey extends readonly string[]> = {
-  all: { kind: 'all' }
-  first: { kind: 'first' }
-  find: {
-    kind: 'find'
-    value: PrimaryKeyInputForRow<row, primaryKey>
-  }
-  count: { kind: 'count' }
-  exists: { kind: 'exists' }
-  insert: {
-    kind: 'insert'
-    values: Partial<row>
-    options?: InsertQueryOptions<row>
-  }
-  insertMany: {
-    kind: 'insertMany'
-    values: Partial<row>[]
-    options?: InsertQueryOptions<row>
-  }
-  update: {
-    kind: 'update'
-    changes: Partial<row>
-    options?: InsertQueryOptions<row>
-  }
-  delete: {
-    kind: 'delete'
-    options?: DeleteQueryOptions<row>
-  }
-  upsert: {
-    kind: 'upsert'
-    values: Partial<row>
-    options?: UpsertQueryOptions<row>
-  }
-}
-
-type QueryExecutionMode = keyof QueryPlanMap<Record<string, unknown>, readonly string[]>
 
 type QueryPhase<
   binding extends QueryBindingState = QueryBindingState,
@@ -124,12 +67,6 @@ type QuerySourceColumnTypes<source extends AnyQuerySource> = QueryColumnTypeMapF
   QuerySourceTableName<source>,
   QuerySourceRow<source>
 >
-
-type QueryPlan<
-  row extends Record<string, unknown>,
-  primaryKey extends readonly string[],
-  mode extends QueryExecutionMode = QueryExecutionMode,
-> = QueryPlanMap<row, primaryKey>[mode]
 
 type QueryResultMap<row extends Record<string, unknown>, loaded extends Record<string, unknown>> = {
   all: Array<row & loaded>
@@ -196,16 +133,6 @@ export type QueryExecutionResult<input> = input extends AnyQuery
 
 type QueryRuntime = {
   exec<input extends AnyQuery>(input: input): Promise<QueryExecutionResult<input>>
-}
-
-type QuerySnapshot<
-  source extends AnyQuerySource = AnyQuerySource,
-  row extends Record<string, unknown> = Record<string, unknown>,
-  mode extends QueryExecutionMode = QueryExecutionMode,
-> = {
-  table: source
-  state: QueryState
-  plan: QueryPlan<row, QuerySourcePrimaryKey<source>, mode>
 }
 
 export const bindQueryRuntime = Symbol('bindQueryRuntime')
@@ -321,14 +248,8 @@ export class Query<
     this: Query<source, columnTypes, row, loaded, QueryAllPhase<phase>>,
     input: WhereInput<QueryColumns<columnTypes>>,
   ): Query<source, columnTypes, row, loaded, QueryAllPhase<phase>> {
-    let predicate = normalizeWhereInput(input)
-    let normalizedPredicate = normalizePredicateValues(
-      predicate,
-      createPredicateColumnResolver([this.#table, ...this.#state.joins.map((join) => join.table)]),
-    )
-
     return this.#clone({
-      where: [...this.#state.where, normalizedPredicate],
+      where: [...this.#state.where, normalizeQueryWhereInput(input, this.#predicateTables())],
     })
   }
 
@@ -336,14 +257,8 @@ export class Query<
     this: Query<source, columnTypes, row, loaded, QueryAllPhase<phase>>,
     input: WhereInput<QueryColumns<columnTypes>>,
   ): Query<source, columnTypes, row, loaded, QueryAllPhase<phase>> {
-    let predicate = normalizeWhereInput(input)
-    let normalizedPredicate = normalizePredicateValues(
-      predicate,
-      createPredicateColumnResolver([this.#table, ...this.#state.joins.map((join) => join.table)]),
-    )
-
     return this.#clone({
-      having: [...this.#state.having, normalizedPredicate],
+      having: [...this.#state.having, normalizeQueryWhereInput(input, this.#predicateTables())],
     })
   }
 
@@ -361,11 +276,7 @@ export class Query<
   > {
     let normalizedOn = normalizePredicateValues(
       on,
-      createPredicateColumnResolver([
-        this.#table,
-        ...this.#state.joins.map((join) => join.table),
-        target,
-      ]),
+      createPredicateColumnResolver([...this.#predicateTables(), target]),
     ) as Predicate<QueryColumns<columnTypes> | QueryColumnName<target>>
 
     return this.#clone({
@@ -600,7 +511,11 @@ export class Query<
   }
 
   [querySnapshot](): QuerySnapshot<source, row, QueryPhaseMode<phase>> {
-    return this.#snapshot()
+    return createQuerySnapshot(
+      this.#table,
+      this.#state,
+      this.#plan as QueryPlan<row, QuerySourcePrimaryKey<source>, QueryPhaseMode<phase>>,
+    )
   }
 
   #resolveTerminal<nextMode extends QueryExecutionMode, result>(
@@ -638,18 +553,7 @@ export class Query<
   #clone(patch: Partial<QueryState>): Query<source, columnTypes, row, loaded, phase> {
     return Query.#createInternal<source, columnTypes, row, loaded, phase>(
       this.#table,
-      {
-        select: patch.select ?? cloneSelection(this.#state.select),
-        distinct: patch.distinct ?? this.#state.distinct,
-        joins: patch.joins ? [...patch.joins] : [...this.#state.joins],
-        where: patch.where ? [...patch.where] : [...this.#state.where],
-        groupBy: patch.groupBy ? [...patch.groupBy] : [...this.#state.groupBy],
-        having: patch.having ? [...patch.having] : [...this.#state.having],
-        orderBy: patch.orderBy ? [...patch.orderBy] : [...this.#state.orderBy],
-        limit: patch.limit === undefined ? this.#state.limit : patch.limit,
-        offset: patch.offset === undefined ? this.#state.offset : patch.offset,
-        with: patch.with ? { ...patch.with } : { ...this.#state.with },
-      },
+      mergeQueryState(this.#state, patch),
       this.#plan as QueryPlan<row, QuerySourcePrimaryKey<source>, QueryPhaseMode<phase>>,
       this.#runtime,
     )
@@ -666,22 +570,16 @@ export class Query<
     )
   }
 
-  #snapshot(): QuerySnapshot<source, row, QueryPhaseMode<phase>> {
-    return {
-      table: this.#table,
-      state: cloneQueryState(this.#state),
-      plan: cloneQueryPlan(
-        this.#plan as QueryPlan<row, QuerySourcePrimaryKey<source>>,
-      ) as QueryPlan<row, QuerySourcePrimaryKey<source>, QueryPhaseMode<phase>>,
-    }
-  }
-
   #boundRuntime(): QueryRuntime {
     if (!this.#runtime) {
       throw new DataTableQueryError('Use db.exec(query) to execute an unbound Query')
     }
 
     return this.#runtime
+  }
+
+  #predicateTables(): AnyTable[] {
+    return [this.#table, ...this.#state.joins.map((join) => join.table)]
   }
 }
 
@@ -707,253 +605,5 @@ export function query<
   >
 }
 
-export function cloneQueryState(state: QueryState): QueryState {
-  return {
-    select: cloneSelection(state.select),
-    distinct: state.distinct,
-    joins: [...state.joins],
-    where: [...state.where],
-    groupBy: [...state.groupBy],
-    having: [...state.having],
-    orderBy: [...state.orderBy],
-    limit: state.limit,
-    offset: state.offset,
-    with: { ...state.with },
-  }
-}
-
-function createInitialQueryState(): QueryState {
-  return {
-    select: '*',
-    distinct: false,
-    joins: [],
-    where: [],
-    groupBy: [],
-    having: [],
-    orderBy: [],
-    with: {},
-  }
-}
-
-function cloneQueryPlan<row extends Record<string, unknown>, primaryKey extends readonly string[]>(
-  plan: QueryPlan<row, primaryKey>,
-): QueryPlan<row, primaryKey> {
-  switch (plan.kind) {
-    case 'all':
-      return { kind: 'all' } as QueryPlan<row, primaryKey>
-    case 'first':
-      return { kind: 'first' } as QueryPlan<row, primaryKey>
-    case 'find':
-      return {
-        kind: 'find',
-        value: clonePrimaryKeyValue(plan.value) as PrimaryKeyInputForRow<row, primaryKey>,
-      } as QueryPlan<row, primaryKey>
-    case 'count':
-      return { kind: 'count' } as QueryPlan<row, primaryKey>
-    case 'exists':
-      return { kind: 'exists' } as QueryPlan<row, primaryKey>
-    case 'insert':
-      return {
-        kind: 'insert',
-        values: { ...plan.values },
-        options: plan.options ? { ...plan.options } : undefined,
-      } as QueryPlan<row, primaryKey>
-    case 'insertMany':
-      return {
-        kind: 'insertMany',
-        values: plan.values.map((value: Partial<row>) => ({ ...value })),
-        options: plan.options ? { ...plan.options } : undefined,
-      } as QueryPlan<row, primaryKey>
-    case 'update':
-      return {
-        kind: 'update',
-        changes: { ...plan.changes },
-        options: plan.options ? { ...plan.options } : undefined,
-      } as QueryPlan<row, primaryKey>
-    case 'delete':
-      return {
-        kind: 'delete',
-        options: plan.options ? { ...plan.options } : undefined,
-      } as QueryPlan<row, primaryKey>
-    case 'upsert':
-      return {
-        kind: 'upsert',
-        values: { ...plan.values },
-        options: plan.options
-          ? {
-              ...plan.options,
-              conflictTarget: plan.options.conflictTarget
-                ? [...plan.options.conflictTarget]
-                : undefined,
-              update: plan.options.update ? { ...plan.options.update } : undefined,
-            }
-          : undefined,
-      } as QueryPlan<row, primaryKey>
-  }
-}
-
-function clonePrimaryKeyValue(value: unknown): unknown {
-  if (typeof value !== 'object' || value === null || Array.isArray(value)) {
-    return value
-  }
-
-  return { ...value }
-}
-
-function cloneSelection(selection: '*' | SelectColumn[]): '*' | SelectColumn[] {
-  if (selection === '*') {
-    return '*'
-  }
-
-  return selection.map((column) => ({ ...column }))
-}
-
-type WriteStatePolicy = {
-  where: boolean
-  orderBy: boolean
-  limit: boolean
-  offset: boolean
-}
-
-function assertWriteState(
-  state: QueryState,
-  operation: 'insert' | 'insertMany' | 'update' | 'delete' | 'upsert',
-  policy: WriteStatePolicy,
-): void {
-  let unsupported: string[] = []
-
-  if (state.select !== '*') unsupported.push('select()')
-  if (state.distinct) unsupported.push('distinct()')
-  if (state.joins.length > 0) unsupported.push('join()')
-  if (state.groupBy.length > 0) unsupported.push('groupBy()')
-  if (state.having.length > 0) unsupported.push('having()')
-  if (Object.keys(state.with).length > 0) unsupported.push('with()')
-  if (!policy.where && state.where.length > 0) unsupported.push('where()')
-  if (!policy.orderBy && state.orderBy.length > 0) unsupported.push('orderBy()')
-  if (!policy.limit && state.limit !== undefined) unsupported.push('limit()')
-  if (!policy.offset && state.offset !== undefined) unsupported.push('offset()')
-
-  if (unsupported.length > 0) {
-    throw new DataTableQueryError(
-      operation + '() does not support these query modifiers: ' + unsupported.join(', '),
-    )
-  }
-}
-
-type ResolvedPredicateColumn = {
-  tableName: string
-  columnName: string
-}
-
-function createPredicateColumnResolver(
-  tables: AnyTable[],
-): (column: string) => ResolvedPredicateColumn {
-  let qualifiedColumns = new Map<string, ResolvedPredicateColumn>()
-  let unqualifiedColumns = new Map<string, ResolvedPredicateColumn>()
-  let ambiguousColumns = new Set<string>()
-
-  for (let table of tables) {
-    let tableColumns = getTableColumns(table)
-    let tableName = getTableName(table)
-
-    for (let columnName in tableColumns) {
-      if (!Object.prototype.hasOwnProperty.call(tableColumns, columnName)) {
-        continue
-      }
-
-      let resolvedColumn: ResolvedPredicateColumn = {
-        tableName,
-        columnName,
-      }
-
-      qualifiedColumns.set(tableName + '.' + columnName, resolvedColumn)
-
-      if (ambiguousColumns.has(columnName)) {
-        continue
-      }
-
-      if (unqualifiedColumns.has(columnName)) {
-        unqualifiedColumns.delete(columnName)
-        ambiguousColumns.add(columnName)
-        continue
-      }
-
-      unqualifiedColumns.set(columnName, resolvedColumn)
-    }
-  }
-
-  return function resolveColumn(column: string): ResolvedPredicateColumn {
-    let qualified = qualifiedColumns.get(column)
-    if (qualified) return qualified
-
-    if (column.includes('.')) {
-      throw new DataTableQueryError('Unknown predicate column "' + column + '"')
-    }
-
-    if (ambiguousColumns.has(column)) {
-      throw new DataTableQueryError(
-        'Ambiguous predicate column "' + column + '". Use a qualified column name',
-      )
-    }
-
-    let unqualified = unqualifiedColumns.get(column)
-
-    if (!unqualified) {
-      throw new DataTableQueryError('Unknown predicate column "' + column + '"')
-    }
-
-    return unqualified
-  }
-}
-
-function normalizePredicateValues(
-  predicate: Predicate,
-  resolveColumn: (column: string) => ResolvedPredicateColumn,
-): Predicate {
-  if (predicate.type === 'comparison') {
-    let column = resolveColumn(predicate.column)
-
-    if (predicate.valueType === 'column') {
-      resolveColumn(predicate.value)
-      return predicate
-    }
-
-    if (predicate.operator === 'in' || predicate.operator === 'notIn') {
-      if (!Array.isArray(predicate.value)) {
-        throw new DataTableValidationError(
-          'Invalid filter value for column "' +
-            column.columnName +
-            '" in table "' +
-            column.tableName +
-            '"',
-          [{ message: 'Expected an array value for "' + predicate.operator + '" predicate' }],
-          {
-            metadata: {
-              table: column.tableName,
-              column: column.columnName,
-            },
-          },
-        )
-      }
-
-      return predicate
-    }
-
-    return predicate
-  }
-
-  if (predicate.type === 'between') {
-    resolveColumn(predicate.column)
-    return predicate
-  }
-
-  if (predicate.type === 'null') {
-    resolveColumn(predicate.column)
-    return predicate
-  }
-
-  return {
-    ...predicate,
-    predicates: predicate.predicates.map((child) => normalizePredicateValues(child, resolveColumn)),
-  }
-}
+export type { QueryState } from './query/state.ts'
+export { cloneQueryState } from './query/state.ts'
