+++ packages/data-table/src/lib/table/ordering.ts
@@ -1,0 +1,12 @@
+/**
+ * Sort direction accepted by `orderBy`.
+ */
+export type OrderDirection = 'asc' | 'desc'
+
+/**
+ * Normalized `orderBy` clause.
+ */
+export type OrderByClause = {
+  column: string
+  direction: OrderDirection
+}
